Handle errors when a tab is preloaded from omnibox
When a tab is being preloaded from the omnibox, it does not really exist and cannot be queries. This generated errors in the extension console. With this fix, we only update the badge for existing tabs and actually remove the data of the fake tab from our store to avoid memory leaks.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>